requirements.txt
Django==4.0.0
djangorestframework==3.14.0
pdfkit==1.0.0
qrcode==7.3.1